Initial Cleanup of builder
